将 HttpClient 换成 HttpWebRequest，完成收藏状态更新（没有删除选项），修改 xaml 格式

+++ Bangumi/Models/BangumiSubject.cs
@@ -13,7 +13,7 @@ namespace Bangumi.Models
         public int id { get; set; }
         public string url { get; set; }
         public int type { get; set; }
-        public int sort { get; set; }
+        public float sort { get; set; }
         public string name { get; set; }
         public string name_cn { get; set; }
         public string duration { get; set; }
